DEV/Insight-Vue > Implement PingPong page to prepare ClientUI for API connection testing
PingPong is a test page that will send a request with axios to a specified
URL.  This one is configured to send to localhost:5000/ping to talk to
the API server.  In the test case, the API server is faked thanks to Sinon's
fake-server functionality.  Issues were found with the synchronous
execution so 'flush-promises' module was added to ensure assertions
were executed after the request returned.